# [wip] Rewrite Array#minmax with Ruby