--- a/array.rb
+++ b/array.rb
@@ -58,4 +58,27 @@ class Array
   def sample(n = (ary = false), random: Random)
     Primitive.rb_ary_sample(random, n, ary)
   end
+
+  # call-seq:
+  #    array.minmax -> [min_val, max_val]
+  #    array.minmax {|a, b| ... } -> [min_val, max_val]
+  #
+  # Returns a new 2-element \Array containing the minimum and maximum values
+  # from +self+, either per method <tt><=></tt> or per a given block:.
+  #
+  # When no block is given, each element in +self+ must respond to method <tt><=></tt>
+  # with an \Integer;
+  # returns a new 2-element \Array containing the minimum and maximum values
+  # from +self+, per method <tt><=></tt>:
+  #    [0, 1, 2].minmax # => [0, 2]
+  #
+  # When a block is given, the block must return an \Integer;
+  # the block is called <tt>self.size-1</tt> times to compare elements;
+  # returns a new 2-element \Array containing the minimum and maximum values
+  # from +self+, per the block:
+  #    ['0', '00', '000'].minmax {|a, b| a.size <=> b.size } # => ["0", "000"]
+  def minmax()
+    Primitive.attr! 'inline'
+    Primitive.cexpr! 'rb_ary_minmax(self)'
+  end
 end